feat(ui): views вынесены в mixins
